requirements.txt
requests==2.25.1
urllib3==1.26.5
Pillow==8.2.0
python-dotenv==0.18.0
imgurpython==1.1.7
argparse==1.4.0